Re-update on the bias adjustment function
Correction to my first edit:
I propose here to start the optimization (choose init values) at the true values for each parameters (excluding the rec devs). Therefore, putting these "new" values of bias ramp-up and ramp-down periods into the "em.ctl" file (in the EM folder) that has already the true values for the init values.

The reason is that for a data poor situation, if we used the "control.ss.new" file, we might be beginning the simulation run from a bad space once we change the rec-dev time series (because the MLE could be biased). As opposed to beginning form the true values which will be on average good.